+++ packages/sdk-communication-layer/src/SocketService.ts
@@ -23,6 +23,7 @@ import { DisconnectOptions } from './types/DisconnectOptions';
 import { KeyInfo } from './types/KeyInfo';
 import { CommunicationLayerLoggingOptions } from './types/LoggingOptions';
 import { logger } from './utils/logger';
+import { RemoteCommunication } from './RemoteCommunication';
 
 export interface SocketServiceProps {
   communicationLayerPreference: CommunicationLayerPreference;
@@ -32,6 +33,7 @@ export interface SocketServiceProps {
   communicationServerUrl: string;
   context: string;
   ecies?: ECIESProps;
+  remote: RemoteCommunication;
   logging?: CommunicationLayerLoggingOptions;
 }
 
@@ -52,6 +54,7 @@ export interface SocketServiceState {
   hasPlaintext: boolean;
   socket?: Socket;
   setupChannelListeners?: boolean;
+  analytics?: boolean;
   keyExchange?: KeyExchange;
 }
 
@@ -86,6 +89,8 @@ export class SocketService extends EventEmitter2 implements CommunicationLayer {
     communicationServerUrl: '',
   };
 
+  remote: RemoteCommunication;
+
   constructor({
     otherPublicKey,
     reconnect,
@@ -94,6 +99,7 @@ export class SocketService extends EventEmitter2 implements CommunicationLayer {
     communicationServerUrl,
     context,
     ecies,
+    remote,
     logging,
   }: SocketServiceProps) {
     super();
@@ -102,6 +108,7 @@ export class SocketService extends EventEmitter2 implements CommunicationLayer {
     this.state.context = context;
     this.state.communicationLayerPreference = communicationLayerPreference;
     this.state.debug = logging?.serviceLayer === true;
+    this.remote = remote;
 
     if (logging?.serviceLayer === true) {
       debug.enable('SocketService:Layer');