+++ packages/sdk-communication-layer/src/services/SocketService/MessageHandlers/handleSendMessage.test.ts
@@ -47,6 +47,12 @@ describe('handleSendMessage', () => {
           areKeysExchanged: jest.fn().mockReturnValue(true),
         },
       },
+      remote: {
+        state: {
+          analytics: true,
+          sdkVersion: '1.0.0',
+        },
+      },
     } as unknown as SocketService;
   });