+++ packages/sdk/src/services/MetaMaskSDK/InitializerManager/setupExtensionPreferences.ts
@@ -41,6 +41,7 @@ export async function setupExtensionPreferences(instance: MetaMaskSDK) {
     try {
       metamaskBrowserExtension = await getBrowserExtension({
         mustBeMetaMask: true,
+        sdkInstance: instance,
       });
 
       window.extension = metamaskBrowserExtension;