+++ packages/sdk-ui/CHANGELOG.md
@@ -6,8 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.0]
+### Added
+- feat: activate sourcemap support in all packages and resolve issues ([#730](https://github.com/MetaMask/metamask-sdk/pull/730))
+- Fix/sdk config provider ([#714](https://github.com/MetaMask/metamask-sdk/pull/714))
+- feat: expose channel id via hook + expose wallet status via sdk instance ([#704](https://github.com/MetaMask/metamask-sdk/pull/704))
+- fix: invalid qrcode url ([#695](https://github.com/MetaMask/metamask-sdk/pull/695))
+- feat: automated github pages publication ([#599](https://github.com/MetaMask/metamask-sdk/pull/599))
+
 ## [0.2.0]
-### Uncategorized
+### Added
 - Revert "Release 48.0.0" ([#686](https://github.com/MetaMask/metamask-sdk/pull/686))
 - Release 48.0.0 ([#684](https://github.com/MetaMask/metamask-sdk/pull/684))
 - feat: permission system flow ([#675](https://github.com/MetaMask/metamask-sdk/pull/675))
@@ -52,7 +60,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - feat: optimize rollup builds ([#496](https://github.com/MetaMask/metamask-sdk/pull/496))
 - fix: linting changelog issue after updating scripts ([#509](https://github.com/MetaMask/metamask-sdk/pull/509))
 
-[Unreleased]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-ui@0.2.0...HEAD
+[Unreleased]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-ui@0.3.0...HEAD
+[0.3.0]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-ui@0.2.0...@metamask/sdk-ui@0.3.0
 [0.2.0]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-ui@0.1.6...@metamask/sdk-ui@0.2.0
 [0.1.6]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-ui@0.1.5...@metamask/sdk-ui@0.1.6
 [0.1.5]: https://github.com/MetaMask/metamask-sdk/compare/@metamask/sdk-ui@0.1.4...@metamask/sdk-ui@0.1.5